Refactored code for aprgbooleanalgebra module

+++ AprgBooleanAlgebra/AprgBooleanAlgebra/src/BooleanAlgebra/Algorithm/QuineMcCluskey/QuineMcCluskey.hpp
@@ -39,19 +39,15 @@ public:
         return result;
     }
 
-    unsigned int getNumberOfOnes(Minterm const value) const {
-        return AlbaBitValueUtilities<Minterm>::getNumberOfOnes(value);
-    }
+    int getNumberOfOnes(Minterm const value) const { return AlbaBitValueUtilities<Minterm>::getNumberOfOnes(value); }
 
-    ImplicantsWithMinterm getImplicants(unsigned int numberOfOnes, unsigned int commonalityCount) const {
+    ImplicantsWithMinterm getImplicants(int numberOfOnes, int commonalityCount) const {
         ImplicantsWithMinterm result;
         auto numberOfOnesIt = m_computationalTable.find(numberOfOnes);
-        if (numberOfOnesIt != m_computationalTable.end()) {
-            ImplicantsMap const& implicantsMap(numberOfOnesIt->second);
+        if (numberOfOnesIt != m_computationalTable.end()) {            ImplicantsMap const& implicantsMap(numberOfOnesIt->second);
             auto commonalityCountIt = implicantsMap.find(commonalityCount);
             if (commonalityCountIt != implicantsMap.end()) {
-                ImplicantsWithMinterm const& implicants(commonalityCountIt->second);
-                for (ImplicantWithMinterm const& implicant : implicants.getImplicantsData()) {
+                ImplicantsWithMinterm const& implicants(commonalityCountIt->second);                for (ImplicantWithMinterm const& implicant : implicants.getImplicantsData()) {
                     result.addImplicant(implicant);
                 }
             }
@@ -73,15 +69,13 @@ public:
         return result;
     }
 
-    bool doImplicantsExistAt(unsigned int numberOfOnes, unsigned int commonalityCount) const {
+    bool doImplicantsExistAt(int numberOfOnes, int commonalityCount) const {
         bool result(false);
         auto numberOfOnesIt = m_computationalTable.find(numberOfOnes);
-        if (numberOfOnesIt != m_computationalTable.end()) {
-            ImplicantsMap const& implicantsMap(numberOfOnesIt->second);
+        if (numberOfOnesIt != m_computationalTable.end()) {            ImplicantsMap const& implicantsMap(numberOfOnesIt->second);
             result = implicantsMap.find(commonalityCount) != implicantsMap.end();
         }
-        return result;
-    }
+        return result;    }
 
     void setInputOutput(Minterm const input, LogicalValue const output) {
         if (output == LogicalValue::True || output == LogicalValue::DontCare) {
@@ -90,19 +84,18 @@ public:
     }
 
     void findAllCombinations() {
-        unsigned int commonalityCount = 0;
+        int commonalityCount = 0;
         bool areAllCombinationsFound(false);
         while (!areAllCombinationsFound) {
             bool isCombinationFound(false);
-            for (unsigned int numberOfOnes = 0; numberOfOnes + 1 < m_computationalTable.size(); numberOfOnes++) {
+            for (int numberOfOnes = 0; numberOfOnes + 1 < static_cast<int>(m_computationalTable.size());
+                 numberOfOnes++) {
                 findCombinationOfImplicants(numberOfOnes, commonalityCount);
                 isCombinationFound = isCombinationFound | doImplicantsExistAt(numberOfOnes, commonalityCount + 1);
-            }
-            areAllCombinationsFound = !isCombinationFound;
+            }            areAllCombinationsFound = !isCombinationFound;
             commonalityCount++;
         }
-        m_maxCommonalityCount = (commonalityCount > 0) ? commonalityCount - 1 : 0;
-    }
+        m_maxCommonalityCount = (commonalityCount > 0) ? commonalityCount - 1 : 0;    }
 
     void fillComputationalTableWithMintermsWithZeroCommonalityCount() {
         for (auto const& inputAndOutputPair : m_inputToOutputMap) {
@@ -110,16 +103,14 @@ public:
         }
     }
 
-    void findCombinationOfImplicants(unsigned int numberOfOnes, unsigned int commonalityCount) {
-        if (numberOfOnes + 1 < m_computationalTable.size()) {
+    void findCombinationOfImplicants(int numberOfOnes, int commonalityCount) {
+        if (numberOfOnes + 1 < static_cast<int>(m_computationalTable.size())) {
             ImplicantsWithMinterm const& implicants1(m_computationalTable[numberOfOnes][commonalityCount]);
             ImplicantsWithMinterm const& implicants2(m_computationalTable[numberOfOnes + 1][commonalityCount]);
-            for (ImplicantWithMinterm const& implicant1 : implicants1.getImplicantsData()) {
-                for (ImplicantWithMinterm const& implicant2 : implicants2.getImplicantsData()) {
+            for (ImplicantWithMinterm const& implicant1 : implicants1.getImplicantsData()) {                for (ImplicantWithMinterm const& implicant2 : implicants2.getImplicantsData()) {
                     if (implicant1.isCompatible(implicant2)) {
                         m_computationalTable[numberOfOnes][commonalityCount + 1].addImplicant(implicant1 + implicant2);
-                    }
-                }
+                    }                }
             }
         }
     }
@@ -142,29 +133,26 @@ public:
         SetOfMinterms inputMintermsWithTrue(getSetOfInputMintermsWithTrue());
 
         while (!inputMintermsWithTrue.empty()) {
-            std::map<Minterm, unsigned int> inputMintermToCountMap;
+            std::map<Minterm, int> inputMintermToCountMap;
             std::map<Minterm, ImplicantWithMinterm> inputMintermToImplicantMap;
-            std::map<unsigned int, ImplicantWithMinterm> countToImplicantMap;
+            std::map<int, ImplicantWithMinterm> countToImplicantMap;
             for (ImplicantWithMinterm const& implicant : finalImplicants.getImplicantsData()) {
-                unsigned int implicantCount(0U);
+                int implicantCount(0U);
                 for (auto const& inputMinterm : inputMintermsWithTrue) {
                     if (implicant.isSuperset(inputMinterm)) {
-                        inputMintermToCountMap[inputMinterm]++;
-                        inputMintermToImplicantMap.emplace(inputMinterm, implicant);
+                        inputMintermToCountMap[inputMinterm]++;                        inputMintermToImplicantMap.emplace(inputMinterm, implicant);
                         implicantCount++;
                     }
                 }
                 countToImplicantMap.emplace(implicantCount, implicant);
             }
-            unsigned int minCount = std::numeric_limits<unsigned int>::max();
+            int minCount = std::numeric_limits<int>::max();
             Minterm mintermWithMinCount;
             for (auto const& inputMintermAndCountPair : inputMintermToCountMap) {
-                if (minCount > inputMintermAndCountPair.second) {
-                    mintermWithMinCount = inputMintermAndCountPair.first;
+                if (minCount > inputMintermAndCountPair.second) {                    mintermWithMinCount = inputMintermAndCountPair.first;
                     minCount = inputMintermAndCountPair.second;
                 }
-            }
-            ImplicantWithMinterm bestFinalImplicant;
+            }            ImplicantWithMinterm bestFinalImplicant;
             if (minCount == 1U) {
                 bestFinalImplicant = inputMintermToImplicantMap.at(mintermWithMinCount);
             } else {
@@ -227,17 +215,16 @@ private:
     }
 
     void addMintermForZeroCommonalityCount(Minterm const& minterm) {
-        unsigned int numberOfOnes(getNumberOfOnes(minterm));
+        int numberOfOnes(getNumberOfOnes(minterm));
         ImplicantWithMinterm implicant;
         implicant.addMinterm(minterm);
         m_computationalTable[numberOfOnes][0].addImplicant(implicant);
     }
 
-    unsigned int m_maxCommonalityCount;
+    int m_maxCommonalityCount;
     InputToOutputMap m_inputToOutputMap;
     ComputationalTable m_computationalTable;  // https://en.wikipedia.org/wiki/Quine%E2%80%93McCluskey_algorithm
 };
-
 }  // namespace booleanAlgebra
 
 }  // namespace alba
